Implement book/unbook and fix session authorization across pages
