+++ app/dashboard/wallet/page.js
@@ -1,214 +1,26 @@
-"use client";
+"use server";
 
-import {
-  Modal,
-  ModalContent,
-  ModalHeader,
-  ModalBody,
-  useDisclosure,
-} from "@nextui-org/modal";
-import {
-  Table,
-  TableHeader,
-  TableBody,
-  TableColumn,
-  TableRow,
-  TableCell,
-} from "@nextui-org/table";
-import { Input } from "@nextui-org/react";
-import { PageTitle, SectionTitle } from "../../components/Titles";
-import { getTransactions } from "../../services/DataService";
-import {
-  inputClassNames,
-  modalClassNames,
-  tableClassNames,
-} from "../../components/ClassNames";
-import { useEffect, useState } from "react";
-import { format } from "date-fns";
+import { redirect } from "next/navigation";
+import { getSession } from "../../lib";
+import WalletPage from "../../components/pages/WalletPage";
 
-export default function Page() {
-  const [searchInput, setSearchInput] = useState("");
-  const [isAdmin, setIsAdmin] = useState(false);
-  const [transactions, setTransactions] = useState([]);
-  const [file, setFile] = useState();
-  const { isOpen, onOpen, onOpenChange } = useDisclosure();
-
-  function formatDate(date) {
-    console.log(`Format data for: ${date}`);
-    return format(date, "d/MM/y HH:mm");
+export default async function Page() {
+  let user;
+  try {
+    const session = await getSession();
+    if (!session) {
+      redirect("/login");
+    }
+    user = {
+      id: session.user.id,
+      name: session.user.name,
+      email: session.user.email,
+      permission: session.user.permission,
+      balance: session.user.balance,
+    }
+  } catch (error) {
+    console.log(error);
+    redirect("/login");
   }
-
-  function creditAccounts() {
-    // TODO: Implement crediting accounts logic
-    return;
-  }
-
-  useEffect(() => {
-    const fetchData = async () => {
-      const transactionsData = await getTransactions();
-      setTransactions(transactionsData);
-    };
-    fetchData();
-  }, []);
-
-  return (
-    <>
-      {isAdmin ? (
-        <div className="h-full flex flex-col gap-y-5">
-          <div className="flex flex-row items-center justify-between">
-            <PageTitle title="Wallet" />
-            <button
-              onClick={onOpen}
-              className="h-[36px] rounded-[30px] px-[20px] bg-[#1F4776] text-white text-sm" // PREVIOUSLY: py-[10px]
-            >
-              Credit accounts
-            </button>
-          </div>
-
-          <div className="h-full flex flex-col rounded-[20px] border border-a-black/10 p-5 bg-white gap-y-2.5">
-            <div className="flex flex-row">
-              <div className="w-3/4">
-                <SectionTitle title="All transactions" />
-              </div>
-              <div className="w-1/4">
-                <Input
-                  placeholder="Search"
-                  value={searchInput}
-                  onValueChange={setSearchInput}
-                  variant="bordered"
-                  size="xs"
-                  classNames={inputClassNames}
-                />
-              </div>
-            </div>
-            <Table removeWrapper classNames={tableClassNames}>
-              <TableHeader>
-                <TableColumn>Date</TableColumn>
-                <TableColumn>Amount</TableColumn>
-                <TableColumn>User</TableColumn>
-                <TableColumn>Description</TableColumn>
-              </TableHeader>
-              <TableBody>
-                {transactions.map((transaction) => {
-                  // TODO: It's not working on getting it in time - Need to join the tables in useEffect()?
-                  // getClassById(transaction.class_id).then((data) => {
-                  //   const className = data.name;
-                  //   fullDescription = `Booked '${className}' (${formatDate(
-                  //     transaction.date
-                  //   )})`;
-                  // });
-                  return (
-                    <TableRow key={transaction.id}>
-                      <TableCell>{formatDate(transaction.date)}</TableCell>
-                      <TableCell>{transaction.amount}</TableCell>
-                      <TableCell>{transaction.user_id}</TableCell>
-                      <TableCell>{transaction.description}</TableCell>
-                    </TableRow>
-                  );
-                })}
-              </TableBody>
-            </Table>
-          </div>
-          <Modal
-            isOpen={isOpen}
-            onOpenChange={onOpenChange}
-            size="md"
-            backdrop="opaque"
-            classNames={modalClassNames}
-          >
-            <ModalContent>
-              {(
-                onClose // TODO: What's the onClose?
-              ) => (
-                <>
-                  <ModalHeader>
-                    <div className="flex flex-row items-center gap-x-2.5">
-                      <SectionTitle title="Credit accounts" />
-                    </div>
-                  </ModalHeader>
-                  <ModalBody>
-                    <div className="md:w-1/3 flex flex-col gap-y-[5px]">
-                      <p className="text-a-black/50 text-sm">CSV file *</p>
-                      <Input
-                        value={file}
-                        onValueChange={setFile}
-                        isRequired
-                        variant="bordered"
-                        size="xs"
-                        classNames={inputClassNames}
-                      />
-                      {/* TODO: Add in upload icon and handle logic */}
-                    </div>
-                    <div className="flex justify-end">
-                      <button
-                        onClick={creditAccounts}
-                        className="rounded-[30px] px-[20px] py-[10px] bg-[#1F4776] text-white"
-                      >
-                        Credit
-                      </button>
-                    </div>
-                  </ModalBody>
-                </>
-              )}
-            </ModalContent>
-          </Modal>
-        </div>
-      ) : (
-        // Start of user Wallet page
-          <div className="h-full flex flex-col gap-y-5">
-          <PageTitle title="Wallet" />
-          <div className="h-1/4 flex flex-row gap-x-5">
-            <div className="w-1/2 rounded-[20px] border border-a-black/10 p-5 bg-white">
-              <p className="font-poppins font-bold text-a-navy text-3xl">10</p>
-              <p className="font-poppins text-[#393E46] text-lg">
-                credits remaining
-              </p>
-            </div>
-            <div className="w-1/2 rounded-[20px] border border-a-black/10 p-5 bg-white">
-              <p className="font-poppins font-bold text-a-navy text-3xl">8</p>
-              <p className="font-poppins text-a-black text-lg">credits spent</p>
-            </div>
-          </div>
-          <div className="h-full flex flex-col rounded-[20px] border border-a-black/10 p-5 bg-white gap-y-2.5">
-            <div className="flex flex-row">
-              <div className="w-3/4">
-                <SectionTitle title="All transactions" />
-              </div>
-              <div className="w-1/4">
-                <Input
-                  placeholder="Search"
-                  value={searchInput}
-                  onValueChange={setSearchInput}
-                  variant="bordered"
-                  size="xs"
-                  classNames={inputClassNames}
-                />
-              </div>
-            </div>
-            <Table removeWrapper classNames={tableClassNames}>
-              <TableHeader>
-                <TableColumn>Date</TableColumn>
-                <TableColumn>Amount</TableColumn>
-                <TableColumn>User</TableColumn>
-                <TableColumn>Description</TableColumn>
-              </TableHeader>
-              <TableBody>
-                {transactions.map((transaction) => {
-                  return (
-                    <TableRow key={transaction.id}>
-                      <TableCell>{format(transaction.date, "d/MM/yyy HH:mm")}</TableCell>
-                      <TableCell>{transaction.amount}</TableCell>
-                      <TableCell>{transaction.user}</TableCell>
-                      <TableCell>{transaction.description}</TableCell>
-                    </TableRow>
-                  );
-                })}
-              </TableBody>
-            </Table>
-          </div>
-        </div>
-        // End of user Wallet page
-      )}
-    </>
-  );
+  return <WalletPage user={user} />;
 }
