+++ app/lib.js
@@ -1,14 +1,16 @@
 import { SignJWT, jwtVerify } from "jose";
 import { cookies } from "next/headers";
 import { createSecretKey } from "crypto";
+import { NextResponse } from "next/server";
 
 const secretKey = createSecretKey(process.env.TOKEN_SECRET, "utf-8");
 
 export async function encrypt(payload) {
+  console.log(payload);
   return await new SignJWT(payload)
     .setProtectedHeader({ alg: "HS256" })
     .setIssuedAt()
-    .setExpirationTime("10 sec from now")
+    .setExpirationTime("1 day from now")
     .sign(secretKey);
 }
 
@@ -21,8 +23,6 @@ export async function decrypt(input) {
 
 export async function getSession() {
   const session = cookies().get("session").value;
-  if (!session) {
-    return null;
-  }
+  if (!session) return null;
   return await decrypt(session);
 }
